Move filter function to filters/
